requirements.txt
colorthief
Pillow